\name{print.cointsmall}
\alias{print.cointsmall}
\title{Print Method for cointsmall Objects}
\usage{
\method{print}{cointsmall}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{"cointsmall"}.}

\item{digits}{Integer. Number of digits to display. Default is \code{4}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for cointsmall Objects
}
