\name{summary.cointsmall}
\alias{summary.cointsmall}
\title{Summary Method for cointsmall Objects}
\usage{
\method{summary}{cointsmall}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"cointsmall"}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary Method for cointsmall Objects
}
